refactor: extract out pairs selection
